made entrypoint script executable and renamed files
