+++ docker-entrypoint.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 set -e
 
-gunicorn --bind=${HOST}:${PORT} broadlink_rm_rest_app:app
+gunicorn --bind=${HOST}:${PORT} rest_app:app
